// src/types/muscles.ts
export type MuscleId =
    | 'chest'
    | 'abs'
    | 'obliques'
    | 'forearms'
    | 'biceps'
    | 'shoulders-front'
    | 'traps'
    | 'lats'
    | 'lower-back'
    | 'glutes'
    | 'hamstrings'
    | 'calves'
    | 'quads'
    | 'triceps'
    | 'shoulders-back'
    | 'head';

export const ALL_MUSCLES: MuscleId[] = [
    'chest',
    'abs',
    'obliques',
    'forearms',
    'biceps',
    'shoulders-front',
    'traps',
    'lats',
    'lower-back',
    'glutes',
    'hamstrings',
    'calves',
    'quads',
    'triceps',
    'shoulders-back',
    'head'
];